requirements.txt
azure-functions
fastapi
uvicorn
motor
pydantic[email]
bcrypt